+++ cs-20151215/alibabacloud_cs20151215/models/_describe_cluster_events_request.py
@@ -7,10 +7,14 @@ from darabonba.model import DaraModel
 class DescribeClusterEventsRequest(DaraModel):
     def __init__(
         self,
+        max_results: int = None,
+        next_token: str = None,
         page_number: int = None,
         page_size: int = None,
         task_id: str = None,
     ):
+        self.max_results = max_results
+        self.next_token = next_token
         # The number of the page to return.
         self.page_number = page_number
         # The number of entries per page. Valid values: 1 to 50. Default value: 50.
@@ -26,6 +30,12 @@ class DescribeClusterEventsRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.max_results is not None:
+            result['max_results'] = self.max_results
+
+        if self.next_token is not None:
+            result['next_token'] = self.next_token
+
         if self.page_number is not None:
             result['page_number'] = self.page_number
 
@@ -39,6 +49,12 @@ class DescribeClusterEventsRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('max_results') is not None:
+            self.max_results = m.get('max_results')
+
+        if m.get('next_token') is not None:
+            self.next_token = m.get('next_token')
+
         if m.get('page_number') is not None:
             self.page_number = m.get('page_number')