+++ cs-20151215/alibabacloud_cs20151215/models/_install_node_pool_components_response_body.py
@@ -11,8 +11,11 @@ class InstallNodePoolComponentsResponseBody(DaraModel):
         request_id: str = None,
         task_id: str = None,
     ):
+        # The ID of the asynchronous task. You can use this ID to track the installation progress.
         self.cluster_id = cluster_id
+        # The unique ID of the request.
         self.request_id = request_id
+        # The task ID.
         self.task_id = task_id
 
     def validate(self):