+++ cs-20151215/alibabacloud_cs20151215/models/_update_kmsencryption_request.py
@@ -10,7 +10,19 @@ class UpdateKMSEncryptionRequest(DaraModel):
         disable_encryption: bool = None,
         kms_key_id: str = None,
     ):
+        # Specifies whether to disable the Secret-at-rest encryption feature.
+        # 
+        # * `true`: Disables the at-rest encryption feature.
+        # * `false`: Enables the at-rest encryption feature.
         self.disable_encryption = disable_encryption
+        # The KMS key ID used by the Secret-at-rest encryption feature.
+        # 
+        # >Notice: You cannot use a service key. You must use either a master key or a customer master key. The key type must be `Aliyun_AES_256`, and the key usage must be `ENCRYPT/DECRYPT`.
+        # 
+        # 
+        # 
+        # 
+        # >Warning: During the process of enabling or disabling the at-rest encryption feature and after the feature is successfully enabled, do not disable or delete the KMS key via the KMS console or OpenAPI. Otherwise, the cluster API Server will become unavailable, preventing normal retrieval of objects such as Secrets and ServiceAccounts, which impacts the normal operation of business applications.
         self.kms_key_id = kms_key_id
 
     def validate(self):