+++ cs-20151215/alibabacloud_cs20151215/models/_update_node_pool_component_request.py
@@ -17,11 +17,17 @@ class UpdateNodePoolComponentRequest(DaraModel):
         rolling_policy: main_models.UpdateNodePoolComponentRequestRollingPolicy = None,
         version: str = None,
     ):
+        # The configuration details for the component update.
         self.config = config
+        # Specifies whether to disable rolling updates. Default: false. If set to false, nodes will be rolled automatically to apply the new configuration.
         self.disable_rolling = disable_rolling
+        # The name of the node component.
         self.name = name
+        # A list of specific nodes to be rolled. If not specified, all nodes in the node pool will be updated.
         self.node_names = node_names
+        # The rolling update configuration.
         self.rolling_policy = rolling_policy
+        # The version of the node component to be updated to.
         self.version = version
 
     def validate(self):
@@ -86,8 +92,11 @@ class UpdateNodePoolComponentRequestRollingPolicy(DaraModel):
         max_parallelism: int = None,
         pause_policy: str = None,
     ):
+        # The time interval between update batches, in seconds.
         self.batch_interval = batch_interval
+        # The maximum number of nodes that can be updated concurrently. Default: 1.
         self.max_parallelism = max_parallelism
+        # The automatic pause strategy during the update process. Valid values: NotPause, FirstBatchPause, EveryBatchPause.
         self.pause_policy = pause_policy
 
     def validate(self):
@@ -127,6 +136,7 @@ class UpdateNodePoolComponentRequestConfig(DaraModel):
         self,
         custom_config: Dict[str, str] = None,
     ):
+        # Custom configuration parameters for the component.
         self.custom_config = custom_config
 
     def validate(self):